# public/openapi.yaml
openapi: 3.0.3
info:
  title: Jotty API
  description: |
    Programmatic access to your checklists and notes with full CRUD operations.

    ## Authentication
    All endpoints require authentication via API key in the `x-api-key` header.

    Get your API key from: Profile → Settings → API Key → Generate
  version: "1.0.0"
  contact:
    name: Jotty
    url: https://github.com/fccview/jotty
security:
  - ApiKeyAuth: []
paths:
...
  /tasks/{taskId}:
    get:
      summary: Get Task
      description: Retrieves a specific task checklist
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
                      category:
                        type: string
                      statuses:
                        type: array
                        items:
                          $ref: '#/components/schemas/KanbanStatus'
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChecklistItem'
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '404':
          description: Task not found
        '400':
          description: Not a task checklist

    put:
      summary: Update Task
      description: Updates task metadata (title, category). All fields are optional.
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Sprint Title"
                category:
                  type: string
                  example: "Work"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
                      category:
                        type: string
                      statuses:
                        type: array
                        items:
                          $ref: '#/components/schemas/KanbanStatus'
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '404':
          description: Task not found
        '400':
          description: Invalid request or not a task checklist

    delete:
      summary: Delete Task
      description: Deletes a task checklist
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task to delete
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          description: Task not found
        '400':
          description: Not a task checklist
...
  /tasks/{taskId}/items:
    post:
      summary: Create Task Item
      description: Creates a new item in a task checklist
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: "Implement user authentication"
                status:
                  type: string
                  default: "todo"
                  example: "todo"
                parentIndex:
                  type: string
                  description: Index path for creating nested items (e.g., "0" for child of first item, "0.1" for grandchild)
                  example: "0"
      responses:
        '200':
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        '400':
          description: Invalid request
        '404':
          description: Task or parent item not found
  /tasks/{taskId}/items/{itemIndex}/status:
    put:
      summary: Update Item Status
      description: Changes an item's status (moves it between Kanban columns)
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task
        - name: itemIndex
          in: path
          required: true
          schema:
            type: string
          description: Index path of the item (e.g., "0", "0.1", "2.0.1")
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  example: "in_progress"
      responses:
        '200':
          description: Item status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Invalid request or item index
        '404':
          description: Task or item not found
  /tasks/{taskId}/items/{itemIndex}:
    delete:
      summary: Delete Task Item
      description: Deletes an item from a task checklist
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the task
        - name: itemIndex
          in: path
          required: true
          schema:
            type: string
          description: Index path of the item to delete (e.g., "0", "0.1", "2.0.1")
      responses:
        '200':
          description: Item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Invalid item index or not a task checklist
        '404':
          description: Task or item not found
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key obtained from Profile → Settings → API Key
  schemas:
...
    KanbanStatus:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the status
          example: "todo"
        label:
          type: string
          description: Display name for the status
          example: "To Do"
        color:
          type: string
          description: Optional color code for the status column
          example: "#3b82f6"
        order:
          type: integer
          description: Display order of the status column
          example: 0
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
          description: Internal item ID
          example: "listId-1234567890"
        index:
          type: integer
          description: Position in the parent's items array
          example: 0
        text:
          type: string
          example: "Task 1"
        completed:
          type: boolean
          example: false
        status:
          type: string
          enum: [todo, in_progress, paused, completed]
          description: Only present for task-type checklists
        time:
          oneOf:
            - type: integer
              example: 0
            - type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  startTime:
                    type: string
                    format: date-time
                  endTime:
                    type: string
                    format: date-time
                  duration:
                    type: integer
          description: Only present for task-type checklists
        children:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
          description: Nested sub-items
...